Moved FHIR handlers into their own package
A little code cleanup to keep things organized.

Also removed a test servlet filter that should never have been committed.

`FHIRRequestFeature` is now injected via Guice.
