--- dpc-web/src/main/java/gov/cms/dpc/web/filters/FHIRServletResponse.java
@@ -1,21 +1,0 @@
-package gov.cms.dpc.web.filters;
-
-import javax.servlet.*;
-import java.io.IOException;
-
-public class FHIRServletResponse implements Filter {
-    @Override
-    public void init(FilterConfig filterConfig) throws ServletException {
-
-    }
-
-    @Override
-    public void doFilter(ServletRequest servletRequest, ServletResponse servletResponse, FilterChain filterChain) throws IOException, ServletException {
-        filterChain.doFilter(servletRequest, servletResponse);
-    }
-
-    @Override
-    public void destroy() {
-
-    }
-}
